requirements.txt
img2table
pandas
google-cloud-aiplatform
camelot-py
openpyxl